+++ apps/marketplace/src/pages/test-modal.tsx
@@ -5,6 +5,7 @@ import Button from '@mui/material/Button';
 import ReportModal from '@/components/modal/ReportModal';
 import MakeOfferModal from '@/components/modal/MakeOfferModal';
 import OnLeaveModal from '@/components/modal/OnLeaveModal';
+import AdvertisementModal from '@/components/marketplace/listing/AdvertisementModal';
 
 const TestModal = () => {
   const report = [
@@ -25,6 +26,11 @@ const TestModal = () => {
   const [openOffer, setOpenOffer] = React.useState(false);
   const [openLeave, setOpenLeave] = React.useState(false);
   const [inputValue, setInputValue] = React.useState<number>(0);
+  const [isOpen, setIsOpen] = React.useState(false);
+
+  const handleClose = (val: boolean) => {
+    setIsOpen(false);
+  };
 
   const router = useRouter();
 
@@ -100,9 +106,15 @@ const TestModal = () => {
         setInputValue={setInputValue}
       />
       <Button onClick={() => setOpenLeave(true)}> On Leave Modal</Button>
-      <OnLeaveModal
-        open={openLeave}
-        setOpen={setOpenLeave}
+      <OnLeaveModal open={openLeave} setOpen={setOpenLeave} />
+      <Button onClick={() => setIsOpen(true)}> Modal</Button>
+      <AdvertisementModal
+        id={1}
+        companyName="Shi lin Fang Metal Comapany 1"
+        description="Are you looking for high-quality metalwork products that are durable and made to last? Look no further than SHI LI FANG IRONWORKS PTE. Our team of skilled craftsmen and engineers use only the best "
+        onClose={handleClose}
+        open={isOpen}
+        url="https://www.google.com"
       />
     </>
   );
